feat: moved more trackig into provider
